Move Azure-related functionality to its own module
